[Fix] #13 - Apply v1.0.1 patch to web experimental branch.

+++ lib/web/interoperability/web_point.dart
@@ -7,7 +7,7 @@ extension type WebPoint._(JSObject _) implements JSObject {
   external double get x;
   external double get y;
 
-  factory WebPoint.fromPoint(KPoint payload) => WebPoint(payload.x, payload.y);
+  factory WebPoint.fromPoint(KPoint payload) => WebPoint(payload.x.toDouble(), payload.y.toDouble());
 
   KPoint toPoint() => KPoint(x, y);
 }
